requirements.txt
# Package management
wheel==0.40.0

# Django
Django==4.2.1

# Webserver
gunicorn==20.1.0

# Database
psycopg2-binary==2.9.6
dj-database-url==2.0.0
sqlparse==0.4.4

# Date/Time
python-dateutil==2.8.2
pytz==2022.7.1
tzdata==2023.3

# Compatibility
six==1.16.0

# APIs
recipe-scrapers==14.36.1
apify-client==1.2.2
spacy==3.5.3
ingredient-parser-nlp==0.1.0b1